requirements.txt
google-cloud-bigquery==3.5.0
functions-framework==3.0.0
requests==2.*
